учёл замечания

+++ задание_4/задача_1.py
@@ -1,10 +1,13 @@
+from sys import argv
+
+
 def salary(hourly_rate:str, hour:str, premium:str) -> str:
     """
     Функция получает параметры зарплаты работника, производит расчёт,
     и возвращает зароботную плату работника 
     """
-    return str((int(hourly_rate)*int(hour))+int(premium))
+    return str(( float(hourly_rate)* float(hour))+ float(premium))
 
-from sys import argv
+
 script_name,hourly_rate, hour, premium = argv
 print(f"расчёт заработной платы сотрудника: {salary(hourly_rate, hour, premium)}")
